Restructure the whole folder
The whole file system was restructured to allow automation of saving
files used with the leetcode extension with vscode. With this new file
structure the workspace for the leetcode extension is changed to the
point to the solution and files are saved in folder in the workspace
named after the language they were written in.
language they were written in.